% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpc_subclasses}
\alias{cpc_subclasses}
\title{data frame of unique CPC subclasses}
\format{
\describe{
\item{cpc_subclass}{cpc subclass}
\item{title}{cpc subclass title}
}
}
\usage{
data(cpc_subclasses)
}
\description{
cpc_subclasses is a data frame of unique CPC subclasses found in \link{cpc_subgroups}
It's the distinct(cpc_subclass,title) where the cpc_subgroup is "00" and
the cpc_group is either "1" or "10"
}
\examples{
data(cpc_subclasses)
head(cpc_subclasses)
}
\keyword{datasets}
