% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpc_subgroups}
\alias{cpc_subgroups}
\title{data frame of CPC classifications}
\format{
\describe{
\item{id}{cpc subclass}
\item{title}{cpc subgroup title}
\item{cpc_subclass}{cpc subclass}
\item{cpc_group}{cpc group}
\item{cpc_subgroup}{cpc subgroup}
}
}
\source{
Data loaded from g_cpc_title.tsv.zip in the \href{patentsview.org/download/data-download-tables}{Patentsview API's bulk data files}. Acknowledged here per its Creative Commons Attribution 4.0 License.
All of the rows are used but the columns have been manipulated as can be seen in data-raw/get_cpc_data.R
}
\usage{
data(cpc_subgroups)
}
\description{
cpc_subgroups is a custom data frame of CPC classifications built from the bulk file
acknowledged below.
}
\examples{
data(cpc_subgroups)
head(cpc_subgroups)
}
\keyword{datasets}
