% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_post.R
\name{pv_post}
\alias{pv_post}
\title{Posts to the Patentsview API by CPC}
\usage{
pv_post(start_date, cpc, env = parent.frame(), ...)
}
\arguments{
\item{start_date}{}
}
\value{
what we got back from the patentsview API
}
\description{
This function formats and submits a POST call to the PatentsView Patent API endpoint
}
\examples{
result <- pv_post("2024-01-01", cpc="G16H")
colnames(result)
}
