% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_patents.R
\name{clean_patents}
\alias{clean_patents}
\title{Queries PatentsView API by CPC}
\usage{
clean_patents(pvresult)
}
\arguments{
\item{pvresult}{}
}
\value{
a data frame of 27 fields
}
\description{
This function formats the returned JSON objects into flat data frames
}
\examples{
clean_patents(pvresult)

}
