% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patents_view.R
\name{patents_view}
\alias{patents_view}
\title{Queries PatentsView API by CPC}
\usage{
patents_view(cpc, from)
}
\arguments{
\item{cpc}{}
}
\value{
a data frame of 27 fields
}
\description{
This function submits requests to the PatentsView API to return information on patents by CPC code. It calls both pv_post() and clean_patents().
}
\examples{
patents_view(cpc="G16H")

}
