% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_post.R
\name{pv_post}
\alias{pv_post}
\title{Queries PatentsView API by CPC}
\usage{
pv_post(page, start_date, env = parent.frame(), ...)
}
\arguments{
\item{page}{}
}
\value{
a data frame of 27 fields
}
\description{
This function formats and submits a POST call to the PatentsView Patents API endpoint
}
\examples{
clean_patents(pvresult)

}
